% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDist.R
\name{fDist}
\alias{fDist}
\title{Functional L2 Distance Between Test and Retest Curves}
\usage{
fDist(data, individuals, test, domain)
}
\arguments{
\item{data}{Data matrix collecting all individuals' curves evaluated on the same grid of \eqn{M} points. 
It is a matrix of dimension \eqn{2n\times M}, where: 
    \eqn{n} is the number of individuals, 
    \eqn{M} is the number of points of the grid.}

\item{individuals}{Factor with \eqn{2n} elements coding the individuals 
(the number of levels correspond to the number of individuals). Note that 
to compute CMC you must have \eqn{n} individuals, and two curves for each individual (corresponding to test and retest).}

\item{test}{Factor with \eqn{2n} elements coding the test sessions 
(the number of levels is 2 corresponding to  test and retest). Note that 
to compute CMC you must have \eqn{2} test sessions for each individual.}

\item{domain}{Vector of dimension \eqn{M} containing the grid points on the x axis where functional data are evaluated.
The grid must be equally spaced.}
}
\value{
A list containing the following components:
\item{dict}{Distance value computed by averaging the values of all individuals.}
\item{dist_subject}{Vector of size \code{n} containing the distance values for each individual. }
}
\description{
Computation of the functional L2 distance for a test-retest
scenario.
}
\examples{


}
\references{
Pini, A., Markström, J., and Schelin, L. (2019): Test–retest reliability measures for curve data: 
an overview with recommendations and supplementary code, \emph{Sports Biomechanics}.
}
\seealso{
See also \code{\link{fICC}} for the functional ICC and \code{\link{fSEM}} for functional SEM.
}
