% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fCMC.R
\name{fCMC}
\alias{fCMC}
\title{Functional Coefficient of Multiple Correlation}
\usage{
fCMC(data, individuals, test)
}
\arguments{
\item{data}{Data matrix collecting all individuals' curves evaluated on the same grid of \eqn{M} points. 
It is a matrix of dimension \eqn{2n\times M}, where: 
    \eqn{n} is the number of individuals, 
    \eqn{M} is the number of points of the grid.}

\item{individuals}{Factor with \eqn{2n} elements coding the individuals 
(the number of levels correspond to the number of individuals). Note that 
to compute CMC you must have \eqn{n} individuals, and two curves for each individual (corresponding to test and retest).}

\item{test}{Factor with \eqn{2n} elements coding the test sessions 
(the number of levels is 2 corresponding to  test and retest). Note that 
to compute CMC you must have \eqn{2} test sessions for each individual.}
}
\value{
A list containing the following components:
\item{CMC}{CMC value computed by averaging the values of all individuals.}
\item{CMC_subject}{Vector of size \code{n} containing the CMC value for each individual. }
}
\description{
Computation of the functional coefficient of multiple correlation (fCMC) for a test-retest
scenario.
}
\details{
Note that the CMC could be undefined for some individuals in the case of curves with a small ROM. 
In such cases, the common CMC value is computed as the average
among all individuals whose CMC is defined.
}
\examples{


}
\references{
Pini, A., Markström, J., and Schelin, L. (2019): Test–retest reliability measures for curve data: 
an overview with recommendations and supplementary code, \emph{Sports Biomechanics}.

Ford, K. R., Myer, G. D., and Hewett, T. E. (2007). Reliability of landing 3D motion analysis: implications for longitudinal analyses. 
\emph{Medicine and Science in Sports and Exercise}, 39, 2021–2028.
}
\seealso{
See also \code{\link{fICC}} for the functional ICC and \code{\link{fSEM}} for functional SEM.
}
